+++ docs/server/mcp/tools/b24-ui-list-documentation-pages.ts
@@ -1,0 +1,22 @@
+import { queryCollection } from '@nuxt/content/server'
+import { withTrailingSlash } from 'ufo'
+
+export default defineMcpTool({
+  title: 'Get Migration Guide',
+  description: 'Lists all documentation pages',
+  cache: '1h',
+  async handler() {
+    const event = useEvent()
+
+    const pages = await queryCollection(event, 'docs').all()
+
+    const config = useRuntimeConfig()
+
+    return jsonResult(pages.map(page => ({
+      title: page.title,
+      description: page.description,
+      path: `${config.public.baseUrl}${withTrailingSlash(page.path)}`,
+      url: `${config.public.canonicalUrl}${config.public.baseUrl}${withTrailingSlash(page.path)}`
+    })))
+  }
+})