+++ docs/server/mcp/tools/b24-ui-list-examples.ts
@@ -1,0 +1,15 @@
+// @ts-expect-error - no types available
+import components from '#component-example/nitro'
+
+export default defineMcpTool({
+  title: 'List Examples',
+  description: 'Lists all available UI examples and code demonstrations',
+  cache: '1h',
+  handler() {
+    const examples = Object.entries<{ pascalName: string }>(components).map(([_key, value]) => {
+      return value.pascalName
+    })
+
+    return jsonResult(examples)
+  }
+})