+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * fix(ColorModeButton): improve icon class merging
 * chore(components): reduce type verbosity by omitting link props from action buttons
 * docs(input-date/input-time/calendar): add note about date format
+* docs(mcp): use `@nuxtjs/mcp-toolkit`
 * chore(vitest): move vue config into vitest project
 
 ## [2.1.8](https://github.com/bitrix24/b24ui/compare/v2.1.7...v2.1.8) (2025-11-25)